test/refactor-reportlog
