+++ testing/pytest.ini
@@ -1,2 +1,3 @@
 [pytest]
 python_files = test_*.py
+log_cli = true
